+++ database/migrations/2020_04_06_173210_create_has_role_table.php
@@ -14,8 +14,8 @@ class CreateHasRoleTable extends Migration
     public function up()
     {
         Schema::create('has_role', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('role_id');
+            $table->unsignedInteger('user_id');
+            $table->unsignedInteger('role_id');
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
